+++ HabboHotel/Catalog/Utilities/ItemUtility.cs
@@ -9,7 +9,7 @@ public static class ItemUtility
     {
         if (!item.Definition.AllowGift || item.IsLimited || item.Amount > 1 || item.Definition.InteractionType == InteractionType.Exchange ||
             item.Definition.InteractionType == InteractionType.Badge || item.Definition.Type != ItemType.Floor && item.Definition.Type != ItemType.Wall || item.CostDiamonds > 0 ||
-            item.Definition.InteractionType == InteractionType.Teleport || item.Definition.InteractionType == InteractionType.Deal)
+            item.Definition.InteractionType == InteractionType.Teleport || item.Definition.IsDeal)
             return false;
         if (item.Definition.IsRare)
             return false;
@@ -21,7 +21,7 @@ public static class ItemUtility
     public static bool CanSelectAmount(CatalogItem item)
     {
         if (item.IsLimited || item.Amount > 1 || item.Definition.InteractionType == InteractionType.Exchange || !item.HaveOffer || item.Definition.InteractionType == InteractionType.Badge ||
-            item.Definition.InteractionType == InteractionType.Deal)
+            item.Definition.IsDeal)
             return false;
         return true;
     }